EllipsoidalGraphEmbedding

Documentation for EllipsoidalGraphEmbedding.

EllipsoidalGraphEmbedding.acc_proj_power_method

EllipsoidalGraphEmbedding.find_community_membership

EllipsoidalGraphEmbedding.partition

EllipsoidalGraphEmbedding.rename_com_unique

EllipsoidalGraphEmbedding.spectral_embed_cluster

EllipsoidalGraphEmbedding.sphere_embed_cluster

EllipsoidalGraphEmbedding.update_centroids

EllipsoidalGraphEmbedding.acc_proj_power_method — Method

acc_proj_power_method(A, X, n_it, tol)

Runs the accelarated projected power iteration on the modularity matrix.

Arguments

A::SparseMatrixCSC{Int64,Int64} sparse graph adjacency matrix.

X::Array{Float64,2} initial matrix.

n_it::Int64 maximum number of iterations.

tol::Float64 tolerance on relative variation of consecutive objective values.

Output

X::Array{Float64,2} approximate stationary matrix of the iteration.

EllipsoidalGraphEmbedding.find_community_membership — Method

find_community_membership(x_embed, R0)

returns the index of the closet centroid to each position vector.

Arguments

x_embed::AbstractArray{Float64,2} array of position vectors

R0::Array{Float64,2}array of position vectors for centroids.

Output

community::Array{Int64,1} membership array

EllipsoidalGraphEmbedding.partition — Method

partition(A, xembed, itmax, n_clusters, p)

returns a clustering of the embedded nodes.

Arguments

A::SparseMatrixCSC{Int64,Int64} adjacency matrix

x_embed::AbstractArray{Float64,2} array of position vectors

it_max::Int64 maximum number of iterations

n_clusters::Int64 maximum number of clusters

p::Array{Float64,1} array of probabilities for sampling the centroids

Output

community::Array{Int64,1} membership array

Q_best::Float64 modularity of the partition

n_updates::Float64 number of centroid updates to reach stationarity

EllipsoidalGraphEmbedding.rename_com_unique — Method

rename_com_unique(community)

rename the community labels by the smallest labels of integers.

Arguments

community::Array{Int64,1} array of integers

Output

community::Array{Int64,1} renamed array of integers

EllipsoidalGraphEmbedding.spectral_embed_cluster — Method

sphere_embed_cluster(A, n_it_PPM, t, n_clusters, n_rep_vec_part, n_updates, shape, r0)

returns a clustering of the embedded nodes.

Arguments

A::SparseMatrixCSC{Int64,Int64} adjacency matrix

it_max::Int64 maximum number of iterations of power method with Gram Schmidt

tol::Float64 tolerance for power method

n_clusters::Int64 maximum number of clusters

n_rep_vec_part::Int64 number of repetitions of vector partition

n_updates::Int64 number of updates of vector partition

dim_embed_spectral::Int64 largest number of dimensions for the embedding

Output

x_embed::AbstractArray{Float64,2} array of position vectors

community::Array{Int64,1} membership array

S::Vector{Float64} eigenvalues spectral embedding

EllipsoidalGraphEmbedding.sphere_embed_cluster — Method

sphere_embed_cluster(A, n_it_PPM, t, n_clusters, n_rep_vec_part, n_updates, shape, r0)

returns a clustering of the embedded nodes.

Arguments

A::SparseMatrixCSC{Int64,Int64} adjacency matrix

n_it_PPM::Int64 maximum number of iterations for project power method (PPM)

tol::Float64 tolerance for relative objective variation for PPM

n_clusters::Int64 maximum number of clusters

n_rep_vec_part::Int64 number of repetitions of vector partition

n_updates::Int64 number of updates of vector partition

shape::String spherical or ellipsoidal embedding

r0::Int64 largest number of dimensions for the embedding

Output

x_embed::AbstractArray{Float64,2} array of position vectors

community::Array{Int64,1} membership array

S::Vector{Float64} singular values embedding

EllipsoidalGraphEmbedding.update_centroids — Method

update_centroids(S, R0, community)

returns a clustering of the embedded nodes.

Arguments

S::AbstractArray{Float64,2} position vectors of embedded nodes

R0::Array{Float64,2} initial position of the centroids

community::Array{Int64,1} initial node membership array

Output

R1::Array{Float64,2} updated position of the centroids

community1::Array{Int64,1} updated node membership array